+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -14,6 +14,7 @@ import io.sentry.hints.Cached
 import io.sentry.hints.DiskFlushNotification
 import io.sentry.hints.TransactionEnd
 import io.sentry.protocol.Contexts
+import io.sentry.protocol.Feedback
 import io.sentry.protocol.Mechanism
 import io.sentry.protocol.Message
 import io.sentry.protocol.Request
@@ -290,6 +291,59 @@ class SentryClientTest {
             SentryLevel.DEBUG
         )
         assertEquals(SentryLevel.DEBUG, sentEvent!!.level)
+    }
+
+    @Test
+    fun `when captureFeedback is called, sentry event contains feedback in contexts and header type`() {
+        var sentEvent: SentryEvent? = null
+        fixture.sentryOptions.setBeforeSend { e, _ -> sentEvent = e; e }
+        val sut = fixture.getSut()
+        val scope = createScope()
+        sut.captureFeedback(Feedback("message"), null, scope)
+
+        val sentFeedback = sentEvent!!.contexts.feedback
+        assertNotNull(sentFeedback)
+        assertEquals("message", sentFeedback.message)
+        assertNull(sentFeedback.replayId)
+        assertNull(sentFeedback.url)
+
+        verify(fixture.transport).send(
+            check {
+                assertEquals(SentryItemType.Feedback, it.items.first().header.type)
+            },
+            anyOrNull()
+        )
+    }
+
+    @Test
+    fun `when captureFeedback, scope replay id is attached to feedback`() {
+        var sentEvent: SentryEvent? = null
+        fixture.sentryOptions.setBeforeSend { e, _ -> sentEvent = e; e }
+        val replayId = SentryId()
+        val sut = fixture.getSut()
+        val scope = createScope()
+        scope.replayId = replayId
+        sut.captureFeedback(Feedback("message"), null, scope)
+
+        val sentFeedback = sentEvent!!.contexts.feedback
+        assertNotNull(sentFeedback)
+        assertEquals(replayId.toString(), sentFeedback.replayId?.toString())
+        assertNull(sentFeedback.url)
+    }
+
+    @Test
+    fun `when captureFeedback, screen is attached to feedback as url`() {
+        var sentEvent: SentryEvent? = null
+        fixture.sentryOptions.setBeforeSend { e, _ -> sentEvent = e; e }
+        val sut = fixture.getSut()
+        val scope = createScope()
+        scope.screen = "screen"
+        sut.captureFeedback(Feedback("message"), null, scope)
+
+        val sentFeedback = sentEvent!!.contexts.feedback
+        assertNotNull(sentFeedback)
+        assertEquals("screen", sentFeedback.url)
+        assertNull(sentFeedback.replayId)
     }
 
     @Test